+++ src/EventGrid/EventGrid/EventSubscription/RemoveAzureEventGridSubscription.cs
@@ -20,7 +20,6 @@ using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 
 namespace Microsoft.Azure.Commands.EventGrid
 {
-    [GenericBreakingChangeWithVersion("Existing syntax will be removed. The newly supported syntax removes the corresponding resource via the parameters [-Name <String>] and [-Scope <String>].", "12.0.0", "2.0.0")]
     [Cmdlet(
         "Remove",
         ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "EventGridSubscription",
@@ -30,6 +29,7 @@ namespace Microsoft.Azure.Commands.EventGrid
 
     public class RemoveAzureRmEventGridSubscription : AzureEventGridCmdletBase
     {
+        [CmdletParameterBreakingChangeWithVersion("ResourceId", "12.0.0", "2.0.0", ChangeDescription = "This parameter will be deprecated.")]
         [Parameter(
             Mandatory = true,
             ValueFromPipelineByPropertyName = true,
@@ -48,6 +48,7 @@ namespace Microsoft.Azure.Commands.EventGrid
         [ValidateNotNullOrEmpty]
         public PSTopic InputObject { get; set; }
 
+        [CmdletParameterBreakingChangeWithVersion("DomainInputObject", "12.0.0", "2.0.0", ChangeDescription = "This parameter will be deprecated.")]
         [Parameter(
             Mandatory = true,
             ValueFromPipeline = true,
@@ -57,6 +58,7 @@ namespace Microsoft.Azure.Commands.EventGrid
         [ValidateNotNullOrEmpty]
         public PSDomain DomainInputObject { get; set; }
 
+        [CmdletParameterBreakingChangeWithVersion("DomainTopicInputObject", "12.0.0", "2.0.0", ChangeDescription = "This parameter will be deprecated.")]
         [Parameter(
             Mandatory = true,
             ValueFromPipeline = true,
@@ -124,6 +126,7 @@ namespace Microsoft.Azure.Commands.EventGrid
         [ValidateNotNullOrEmpty]
         public string EventSubscriptionName { get; set; }
 
+        [CmdletParameterBreakingChangeWithVersion("ResourceGroupName", "12.0.0", "2.0.0", ChangeDescription = "This parameter will be deprecated.")]
         [Parameter(
             Mandatory = true,
             ValueFromPipelineByPropertyName = true,
@@ -147,6 +150,7 @@ namespace Microsoft.Azure.Commands.EventGrid
         [Alias(AliasResourceGroup)]
         public string ResourceGroupName { get; set; }
 
+        [CmdletParameterBreakingChangeWithVersion("TopicName", "12.0.0", "2.0.0", ChangeDescription = "This parameter will be deprecated.")]
         [Parameter(
             Mandatory = true,
             ValueFromPipelineByPropertyName = true,
@@ -156,6 +160,7 @@ namespace Microsoft.Azure.Commands.EventGrid
         [ValidateNotNullOrEmpty]
         public string TopicName { get; set; }
 
+        [CmdletParameterBreakingChangeWithVersion("DomainName", "12.0.0", "2.0.0", ChangeDescription = "This parameter will be deprecated.")]
         [Parameter(
             Mandatory = true,
             ValueFromPipelineByPropertyName = true,
@@ -165,6 +170,7 @@ namespace Microsoft.Azure.Commands.EventGrid
         [ValidateNotNullOrEmpty]
         public string DomainName { get; set; }
 
+        [CmdletParameterBreakingChangeWithVersion("DomainTopicName", "12.0.0", "2.0.0", ChangeDescription = "This parameter will be deprecated.")]
         [Parameter(
             Mandatory = false,
             ValueFromPipelineByPropertyName = true,